PP-5428 Merge GoCardlessServiceOnDemandTest with GoCardlessServiceTest
